% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locStra.r
\name{makeWindows}
\alias{makeWindows}
\title{Auxiliary function to generate a two-column matrix of windows to be used in the function 'fullscan'.}
\usage{
makeWindows(len, size, offset)
}
\arguments{
\item{len}{The overall length of the data which is to be scanned in windows.}

\item{size}{The window size.}

\item{offset}{The offset of the generated windows (e.g., if \code{offset=1} then sliding window, if \code{offset=size} then blocks).}
}
\value{
A two-column matrix of sliding windows, with one window per row defined through start and end value.
}
\description{
Auxiliary function to generate a two-column matrix of windows to be used in the function 'fullscan'.
}
\examples{
require(locStra)
print(makeWindows(100,10,5))

}
